Start of development
This test-suite is designed to evaluate the timing behaviour of the
pcl/people-tracker library. To this end, the components of this test
suite and their functionalities are realised as described below.

class testSetup:
An instance of this class handles all inputs necessary to perform a
test session. There are inputs read from the configuration file
TestConfig.txt, which are supposed to remain unchanged among a batch
of tests. Parameters having impact on the timing and result quality
as well as references to input and output files (stream of pointclouds,
reference date, output files) are parsed from the command line.

class testExecution:
An instance of this class handles the execution of a test session and
the collection of associated data. A rosbag file containing a pointcloud
stream serves as input. Each frame is evaluated separately. The resulting
dataset yields the execution time and information regarding detected
objects for each frame.

class testEval:
An instance of this class handles the evaluation of test results.
To this end, a reference file containing ground truth data is employed.
A comparison between test results and ground truth data evaluates the
accuracy of the computation. The results are stored within two files.
One indicates the position of detected object, the other one provides
information regarding the execution time and the quality of results.
The latter can be used to generate plots using gnuplot.

	neue Datei:             src/main.cpp