Refactor code towards distinct separation of concerns
The test suite is supposed to separate the initialisation,
execution and result evaluation issues. This has not been
implemented consistently, which is indicated by the use of many
get methods that only pass data.

In order to establish a clear separation of concerns, the get
methods are replaced by initialisation methods. The outcome is a
distinct encapsulation of the initialisation of data and objects
employed within the execution and evaluation part.