Fix cross-compilation issues
During the build process for the target, implicitly included
libraries turned out to be missing. To resolve these dependencies,
the C++ standard libraries, fstream, ios and iomanip, have been
included.

Additionally, the dependencies regarding ROS needed to be noted as
well as the install directive, since this has been considered on the
development system implicitly.