Introduce real-time features
The survey of people library's impact on the timing behaviour of
real-time applications depending on its parametrization requires the
preclusion of interference arising from non-rt-scheduling and page
faults. To this end, following provisions have been made:

 - Lock virtual address space into RAM using mlockall. This prevents
   the memory from being paged to the swap area.

 - Claim the test process being real time capable. For that purpose
   either RR or FIFO scheduling policy must be applied. In addition,
   the priority value must be assigned.

Note that it is not recommended to assign the application's priority
to the highest value of 99, since there are a few management threads
that need to run with higher priority than the application.