Shift minc option to test configuration
The minc option serves the purpose to set the threshold for the HOG
confidence. That is, clusters are only considered a match that exhibit
a confidence value greater than minc.

As opposed to the other options this one is of no interest regarding
the impact on the execution time. Hence it makes the usage of the test
suite more convenient by removing this option from the command line
prompt.

However, the HOG confidence affects the accuracy of detection. Since
there is no fixed value minc is to be assigned to the capability to
set this parameter has been shifted to the configuration file.